% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_button.R
\name{dynamic_buttonInput}
\alias{dynamic_buttonInput}
\title{Dynamic Button}
\usage{
dynamic_buttonInput(inputId, maxHeight = "3.5")
}
\description{
React Button for multi-wallet connect using Dynamic SDK
}
