% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_button.R
\name{updateDynamic_buttonInput}
\alias{updateDynamic_buttonInput}
\title{Dyanamic Button}
\usage{
updateDynamic_buttonInput(session, inputId, value, configuration = NULL)
}
\description{
React Button for multi-wallet connect using Dynamic SDK
}
